+++ benchmarks/benchmarks.py
@@ -19,7 +19,7 @@ from tdastro.models.basic_models import ConstantSEDModel, LinearWavelengthModel,
 from tdastro.models.lightcurve_template_model import LightcurveTemplateModel
 from tdastro.models.multi_object_model import AdditiveMultiObjectModel
 from tdastro.models.sncomso_models import SncosmoWrapperModel
-from tdastro.models.static_sed_source import StaticSEDModel
+from tdastro.models.static_sed_model import StaticSEDModel
 
 # ASV runs from copy of the project (benchmarks/env/....). So we load the
 # data files based off the current file location instead.
