+++ src/tdastro/models/static_sed_model.py
