+++ tests/tdastro/models/test_multi_object_models.py
@@ -5,7 +5,7 @@ from tdastro.effects.basic_effects import ConstantDimming
 from tdastro.math_nodes.np_random import NumpyRandomFunc
 from tdastro.models.basic_models import ConstantSEDModel, StepModel
 from tdastro.models.multi_object_model import AdditiveMultiObjectModel, RandomMultiObjectModel
-from tdastro.models.static_sed_source import StaticBandfluxModel, StaticSEDModel
+from tdastro.models.static_sed_model import StaticBandfluxModel, StaticSEDModel
 
 
 def test_additive_multi_object_node() -> None:
